Move `SMW\Page` to `SMW\MediaWiki\Page`
